If not running interactively, don't do anything
This helped me in removing a crucial bug where scp was not working. All scp was
doing was giving me a bunch of underscores. This was due to the line

fortune | cowsay

Due to its output, scp gets confused. And throws the error. The check for
interactive console eliminates this.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>